iwyu: algorithm
Move CanCompareIndices implementation
